Add Meson run target for running uninstalled
We determined in a discussion in the #javascript channel that when using
Meson, the goal of running the program uninstalled with no special
environment variables is no longer feasible. However, Meson does provide
"run targets", which allow you to run a program after having built certain
dependencies.

We add a custom target to compile the GSettings schema locally in the
build dir, and a run target named "devel" to run the program uninstalled.
The "devel" target depends on building the GResources and compiling the
GSettings schema, so it should reflect any changes that are made in the
source during development.

Closes: #15